requirements.txt
streamlit
python-pptx
openai
requests
pywin32
python-dotenv 